requirements.txt
flask==3.1.1
gspread==6.2.1
line-bot-sdk==3.3.0